Add files via upload
HTML, or Hypertext Markup Language, is the basic code used to create web pages. It uses tags to mark different types of content like headings, paragraphs, images, and links. These tags define how content is structured and displayed on websites. HTML documents are made up of nested elements, each with its own purpose and settings. It's essential for building websites because it lays the foundation for combining text, images, and other media to create the pages you see online.